# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "version": "1.0.0",
      "title": "Todo-App Backend API",
      "description": "This is a simple CRUD API for managing todos in an e-commerce platform.",
      "termsOfService": "",
      "contact": {
        "name": "API Support",
        "url": "http://www.example.com/support",
        "email": "support@example.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local server"
      }
    ],
    "tags": [
      {
        "name": "Todos",
        "description": "API for Todo operations"
      },
      {
        "name": "Users",
        "description": "API for Users in the system"
      }
    ],
    "paths": {
      "/todos": {
        "get": {
          "tags": ["todos"],
          "summary": "Get all Todos",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/todos"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["todos"],
          "summary": "Create new todo",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/todos"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/todos"
                  }
                }
              }
            }
          }
        }
      },
      "/todos/{id}": {
        "get": {
          "tags": ["todos"],
          "summary": "Get a todo by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of todo to retrieve",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/todos"
                  }
                }
              }
            }
          }
        },
        "patch": {
          "tags": ["todos"],
          "summary": "Update a todo by ID",
          "parameters": [
            {
              "name": "todoId",
              "in": "path",
              "required": true,
              "description": "ID of todo to update",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/todos"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/todos"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["todos"],
          "summary": "Delete a todo by ID",
          "parameters": [
            {
              "name": "todoId",
              "in": "path",
              "required": true,
              "description": "ID of todo to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/user/login": {
        "post": {
          "tags": ["Users"],
          "summary": "User Login",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              },
              "required": true
            }
          },
          "responses": {
            "201": {
              "description": "User Logged in successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Users"
                  }
                }
              }
            }
          }
        }
      },
      "/user/register": {      
        "post": {
          "tags": ["Users"],
          "summary": "Register a new User",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              },
              "required": true
            }
          },
          "responses": {
            "201": {
              "description": "User registered successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Users"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "todos": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "relatedType" : {
              "$ref": "#/components/schemas/todos"
            }
          }
        },
        "Users": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "username": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "relatedType" : {
              "$ref": "#/components/schemas/Users"
            }
          }
        }
      }
    }
  }